Moving Attributes and Gender from Entities to ValueObjects folder.
