+++ app/Modules/Character/Domain/ValueObjects/Attributes.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace App\Modules\Character\Domain\Entities;
+namespace App\Modules\Character\Domain\ValueObjects;
 
 
 use Illuminate\Support\Collection;
